@@ -11,17 +11,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.2.0",
+    "@angular/animations": "^14.0.6",
     "@angular/cdk": "^13.2.6",
-    "@angular/common": "~13.2.0",
-    "@angular/compiler": "~13.2.0",
-    "@angular/core": "~13.2.0",
+    "@angular/common": "^14.0.6",
+    "@angular/compiler": "^14.0.6",
+    "@angular/core": "^14.0.6",
     "@angular/flex-layout": "^13.0.0-beta.38",
-    "@angular/forms": "~13.2.0",
+    "@angular/forms": "^14.0.6",
     "@angular/material": "^13.2.6",
-    "@angular/platform-browser": "~13.2.0",
-    "@angular/platform-browser-dynamic": "~13.2.0",
-    "@angular/router": "~13.2.0",
+    "@angular/platform-browser": "^14.0.6",
+    "@angular/platform-browser-dynamic": "^14.0.6",
+    "@angular/router": "^14.0.6",
     "apexcharts": "^3.35.3",
     "chart.js": "^3.7.1",
     "chartjs-plugin-datalabels": "^2.0.0",
@@ -37,9 +37,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.0",
-    "@angular/cli": "~13.2.0",
-    "@angular/compiler-cli": "~13.2.0",
+    "@angular-devkit/build-angular": "^14.0.6",
+    "@angular/cli": "^14.0.6",
+    "@angular/compiler-cli": "^14.0.6",
     "@types/html2canvas": "^1.0.0",
     "@types/jasmine": "~3.10.0",
     "@types/jspdf": "^2.0.0",
@@ -50,7 +50,7 @@
     "karma-coverage": "~2.1.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.5.2"
+    "typescript": "~4.7.4"
   },
   "engines": {
     "node": "16.13.1",
